+++ e2e-tests/insurance/cypress/e2e/journeys/policy/broker/validation/broker-postcode-validation.spec.js
@@ -47,7 +47,7 @@ context('Insurance - Policy - Broker Page - Validation - Postcode', () => {
 
       cy.completeAndSubmitPolicyTypeForm(FIELD_VALUES.POLICY_TYPE.SINGLE);
       cy.completeAndSubmitSingleContractPolicyForm({});
-      cy.completeAndSubmitAboutGoodsOrServicesForm();
+      cy.completeAndSubmitAboutGoodsOrServicesForm({});
       cy.completeAndSubmitNameOnPolicyForm({ sameName: true });
 
       url = `${Cypress.config('baseUrl')}${ROOT}/${referenceNumber}${BROKER_ROOT}`;