+++ e2e-tests/quote/cypress/e2e/journeys/cookies-consent/change-cookies-consent-via-banner-and-cookies-page.spec.js
@@ -41,7 +41,7 @@ context('Cookies consent - change via banner and cookies page', () => {
       cy.checkCookiesConsentBannerDoesNotExist();
     });
 
-    it('should render a google tag manager script and data layer script', () => {
+    it('should render a google analytics and google tag manager scripts', () => {
       cy.checkAnalyticsScriptsAreRendered();
     });