+++ src/api/constants/application/versions/index.ts
@@ -18,6 +18,12 @@ const VERSIONS = [
     DEFAULT_FINAL_DESTINATION_KNOWN: true,
     DEFAULT_NEED_PRE_CREDIT_PERIOD_COVER: false,
   },
+  {
+    VERSION_NUMBER: '2',
+    OVER_500K_SUPPORT: true,
+    DEFAULT_FINAL_DESTINATION_KNOWN: null,
+    DEFAULT_NEED_PRE_CREDIT_PERIOD_COVER: null,
+  },
 ] as Array<ApplicationVersion>;
 
 export default VERSIONS;