+++ src/ui/server/test-mocks/index.ts
@@ -23,7 +23,7 @@ import mockContact from './mock-contact';
 import { PRODUCT } from '../content-strings';
 import { INTEGRITY } from '../constants';
 
-const { JS, GOVUK, FORM, COOKIES, GA, MOJ, ACCESSIBILITY } = INTEGRITY;
+const { JS, GOVUK, FORM, COOKIES, GA, GA_TAG_MANAGER, MOJ, ACCESSIBILITY } = INTEGRITY;
 
 const mockReq = () => {
   const req = {
@@ -81,6 +81,7 @@ const mockRes = () => {
       FORM,
       COOKIES,
       GA,
+      GA_TAG_MANAGER,
     },
   };